Page-Agile/Workspace-reservation: Pressable Component
