+++ src/screens/agileOffice/screens/AgileOfficeWorkspaceReservation/components/index.tsx
@@ -1,1 +1,2 @@
 export { default as Select } from "./Select";
+export { default as Pressable } from "./Pressable";
