+++ src/screens/agileOffice/screens/AgileOfficeWorkspaceReservation/index.tsx
@@ -1,5 +1,5 @@
 import React from "react";
-import { Platform, Pressable, StyleSheet } from "react-native";
+import { Platform, StyleSheet } from "react-native";
 import moment from "moment";
 import { FontAwesome5 } from "@expo/vector-icons";
 import DateTimePicker from "@react-native-community/datetimepicker";
@@ -19,7 +19,7 @@ import {
 import Button from "components/Button";
 import { ScreenWrapper, unifi_c1, unifi_c4, unifi_c9 } from "components/styles";
 
-import { Select } from "./components";
+import { Select, Pressable } from "./components";
 
 import useWorkspaceReservation from "./useWorkspaceReservation";
 
@@ -41,23 +41,6 @@ export function AgileOfficeWorkspaceReservation({ navigation }) {
   } = useWorkspaceReservation(navigation);
   const textColor = useColorModeValue("black", unifi_c9);
   const placeholderTextColor = useColorModeValue("#C2C2C2", unifi_c9);
-  const selectBorderColor = useColorModeValue("#707070", undefined);
-  const selectBackgroundColor = useColorModeValue("white", "#707070");
-  const dateTimeSelectStyle = (type: string): {} => ({
-    flexDirection: "row",
-    justifyContent: "space-between",
-    width: type === "full" ? "100%" : "50%",
-    alignItems: "center",
-    backgroundColor: selectBackgroundColor,
-    padding: 10,
-    borderColor: selectBorderColor,
-    borderWidth: 1,
-    borderLeftWidth: type === "right" ? 0 : undefined,
-    borderTopLeftRadius: type === "right" ? 0 : 6,
-    borderBottomLeftRadius: type === "right" ? 0 : 6,
-    borderTopRightRadius: type === "left" ? 0 : 6,
-    borderBottomRightRadius: type === "left" ? 0 : 6,
-  });
 
   const placeholderStyle = { fontSize: 16, color: placeholderTextColor };
   const textStyle = { fontSize: 16, color: textColor };
@@ -143,7 +126,7 @@ export function AgileOfficeWorkspaceReservation({ navigation }) {
         </HStack>
         <HStack my={3} alignItems="flex-start">
           <Pressable
-            style={dateTimeSelectStyle(isMoreThan1Day ? "left" : "full")}
+            type={isMoreThan1Day ? "left" : "full"}
             onPress={() => {
               setTempDateTime(fromDate || new Date());
               setDateTimePickerConfig({ mode: "date", onChange: setFromDate });
@@ -160,7 +143,7 @@ export function AgileOfficeWorkspaceReservation({ navigation }) {
           </Pressable>
           {isMoreThan1Day && (
             <Pressable
-              style={dateTimeSelectStyle("right")}
+              type="right"
               onPress={() => {
                 setTempDateTime(toDate || new Date());
                 setDateTimePickerConfig({ mode: "date", onChange: setToDate });
@@ -176,7 +159,7 @@ export function AgileOfficeWorkspaceReservation({ navigation }) {
         <Text style={styles.label}>Choose Time</Text>
         <HStack my={3} alignItems="flex-start">
           <Pressable
-            style={dateTimeSelectStyle("left")}
+            type="left"
             onPress={() => {
               setTempDateTime(time.from || new Date());
               setDateTimePickerConfig({
@@ -191,7 +174,7 @@ export function AgileOfficeWorkspaceReservation({ navigation }) {
             <Icon as={FontAwesome5} name="chevron-down" size={4} />
           </Pressable>
           <Pressable
-            style={dateTimeSelectStyle("right")}
+            type="right"
             onPress={() => {
               setTempDateTime(time.to || new Date());
               setDateTimePickerConfig({
